Include Tag Dimension for Utilization
Each process includes a tag attribute. This is by default the name of
the folder Sidekiq is running in, but can be configured in sidekiq.yml
by setting `tag: cluster-one`.

Since metrics are emited from the Sidekiq Enterprise leader identifying the
utilization of a subset of processes is not possible via the additional
dimensions feature.